Space Invaders game
A traditional space invaders game with 5 levels and an end of game boss
to defeat. The player controls the spaceship and shoots at the aliens
with a laser. Each alien hit will increase the score by 1. The score
resets with every level but the health of the player remains as it is
until the end of game. Meanwhile health and boost powerups will be
popping up throughout the game to help the user. If the player loses the
game, he is redirected to the main menu to play the game again.